Create IO panel with layout and no custom controls
The IO panel is laid out as a table, with rows for inputs (analogue and
digital) and outputs. Where "[ ]" is displayed currently, a switch icon
will be drawn indicating the state of the output, to emphasize that it
can be changed.

Both digital inputs and outputs will be highlighted when on. Analogue
inputs will have vertical bars behind the numbers, showing the current
value.

For testing, any panel created is now an IO panel.