Distinguish between set and unset servos
If a servo has not been set since the board was powered up, its slider
will appear orange and be set to 50. To set that servo, the user must
change the slider to the desired value and press the rotary encoder.
This is to prevent mechanical damage to the servo or the mechanism to
which it is connected.